Extract out common functions for future use
