+++ js-exercises/array-utils/common/isCallable.js
@@ -1,0 +1,7 @@
+const isCallable = (argument) => {
+  let value = false;
+  value = (typeof argument === 'function');
+  return value;
+};
+
+module.exports = isCallable;
