+++ js-exercises/array-utils/custom-for-each.js
@@ -1,32 +1,7 @@
 
-const ToObject = (value) => {
-  if (value === null || value === undefined) {
-    throw new TypeError(`${value} is either null or undefined`);
-  }
-
-  // Checks of boolean, string, number, symbol are not present, as they are out of scope
-
-  if (typeof value !== 'object') {
-    throw new TypeError(`${value} is not an array`);
-  }
-  // returning object, which in our case should be an array
-  return value;
-};
-
-const ToLength = (array) => {
-  if (typeof array.length === 'number' && array.length >= 0) {
-    return array.length;
-  }
-
-  // This error is not in specification. But I needed this to validate if it's array or not.
-  throw new TypeError(`${array} is not an array`);
-};
-
-const isCallable = (argument) => {
-  let value = false;
-  value = (typeof argument === 'function');
-  return value;
-};
+import { ToObject } from './common/toObject';
+import { ToLength } from './common/toLength';
+import { isCallable } from './common/isCallable';
 
 const customForEach = (array, callback, thisArgument) => {
   const object = ToObject(array);
